Print the tone generator's phase tolerance portably
TONE_PHASE_TOLERANCE derives from UINT32_MAX, which picolibc spells
with a UL suffix on 32-bit targets, so the %u it was printed with is a
format mismatch there. The pipeline suite builds for native_sim on the
host, where the constant is a plain unsigned int, and for
nucleo_h723zg in the board runs, where -Werror=format turned it into a
build failure once the suite carried this case.

Cast both printed sample counts at the call site rather than widening
the format, so the message reads the same on every target.

+++ tests/subsys/audio/pipeline/test_tone_gen.c
@@ -556,7 +556,8 @@ ZTEST(audio_pipeline_tone_gen, test_source_phase_does_not_drift_over_ten_seconds
 
 	zassert_true(tone_abs64(error) <= (int64_t)TONE_PHASE_TOLERANCE,
 		     "phase is off by %lld of 2^32 after %u samples, tolerance is %u",
-		     (long long)error, TONE_LONG_SAMPLES, TONE_PHASE_TOLERANCE);
+		     (long long)error, (unsigned int)TONE_LONG_SAMPLES,
+		     (unsigned int)TONE_PHASE_TOLERANCE);
 
 	/* The amplitude has to hold up as well: a resonator recurrence keeps
 	 * its frequency far better than its level, and a stimulus that fades
